% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MALDIassist-package.R
\docType{package}
\name{MALDIassist-package}
\alias{MALDIassist}
\alias{MALDIassist-package}
\title{MALDIassist: Mathematical utilities for MALDI-TOF mass spectrometry}
\description{
Tools for loading Bruker MALDI-TOF spectra, preprocessing, Gaussian KDE-based
peak detection, peak-quality metrics, and cohort feature analysis workflows.
}
\section{Main functions}{

\itemize{
\item \code{\link[=load_maldi_spectra]{load_maldi_spectra()}}: load Bruker raw spectra
\item \code{\link[=preprocess_maldi_spectra]{preprocess_maldi_spectra()}}: smooth and baseline-correct spectra
\item \code{\link[=find_peaks]{find_peaks()}} / \code{\link[=find_peaks_spectra]{find_peaks_spectra()}}: detect ordinary and shoulder peaks
in one spectrum or a list of spectra
\item \code{\link[=find_peaks_fast]{find_peaks_fast()}} / \code{\link[=find_peaks_spectra_fast]{find_peaks_spectra_fast()}}: fast local peak detection
in one spectrum or a list of spectra
\item \code{\link[=filter_peaks]{filter_peaks()}} / \code{\link[=filter_peaks_spectra]{filter_peaks_spectra()}}: filter detected peaks by
intensity, prominence, and strength
\item \code{\link[=build_kde_spectrum]{build_kde_spectrum()}} / \code{\link[=build_kde_spectra]{build_kde_spectra()}}: build Gaussian KDE spectra
\item \code{\link[=find_frequent_mz]{find_frequent_mz()}}: find frequent m/z values across a cohort
\item \code{\link[=filter_overlap_mz]{filter_overlap_mz()}}: remove overlapping nearby frequent m/z features
\item \code{\link[=align_spectra]{align_spectra()}}: align spectra to internal standards (linear / lowess)
\item \code{\link[=build_matched_matrix]{build_matched_matrix()}}: assemble a cohort peak intensity matrix
}
}

\section{Suggested packages}{

Some functions require optional packages that are only loaded when used:
\itemize{
\item \code{\link{visualize_spectrum}}, \code{\link{visualize_spectra}}: colorspace
\item \code{\link{heatmap_matched_matrix}}: pheatmap
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hiows/MALDIassist}
  \item Report bugs at \url{https://github.com/hiows/MALDIassist/issues}
}

}
\author{
\strong{Maintainer}: Wonseok Oh \email{hiows97@gmail.com} (\href{https://orcid.org/0009-0002-0687-8466}{ORCID})

Authors:
\itemize{
  \item Wonseok Oh \email{hiows97@gmail.com} (\href{https://orcid.org/0009-0002-0687-8466}{ORCID})
}

}
